parser reorganisation
