+++ src/lib/utility/parsing/parser.ts
@@ -1,0 +1,3 @@
+export interface Parser<T> {
+    parse(text: string): T | undefined;
+}
