+++ src/lib/utility/public-api.ts
@@ -2,11 +2,11 @@ export * from './multiset/public-api';
 export * from './prefix-graphs/public-api';
 export * from './glpk/public-api';
 export * from './svg/public-api';
+export * from './parsing/public-api';
 
 export * from './incrementing-counter';
 export * from './file-reader.service';
 export * from './drop-file';
-export * from './abstract-block-parser';
 export * from './algorithm-result';
 export * from './arraify';
 export * from './relabeler';
